environments/local/aiq_detector/requirements.txt
kserve>=0.15.0
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
pillow>=9.5.0
pydantic>=2.0.0
numpy>=1.24.0
fastapi>=0.100.0